Move all of the execution-based utilities to their own file
